pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Akmaljon701 <akmalkhanever@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django-cleanup = "^8.1.0"
django-cors-headers = "^4.4.0"
django-debug-toolbar = "^4.4.6"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.7.1"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"
django = "^5.1.1"
psycopg2 = "^2.9.9"
redis = "^5.0.8"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
